rename `instantiate` module to `subst`
